+++ src/main/java/ar/com/kfgodel/function/intertype/floats/FloatToFloatFunction.java
@@ -3,15 +3,14 @@ package ar.com.kfgodel.function.intertype.floats;
 import java.util.function.Function;
 
 /**
- * This type represents a function that takes a boolean input and produces a float as result
  * Date: 29/07/17 - 19:57
  */
-public interface FloatToFloatFunction extends Function<Float, Float> {
+public interface FloatToFloatFunction extends Function<Integer, Float> {
 
   float apply(float input);
 
   @Override
-  default Float apply(Float input){
-    return apply(input.floatValue());
+  default Float apply(Integer input){
+    return apply(input.intValue());
   }
 }
