+++ src/main/java/ar/com/kfgodel/function/intertype/floats/FloatToIntFunction.java
@@ -3,15 +3,14 @@ package ar.com.kfgodel.function.intertype.floats;
 import java.util.function.Function;
 
 /**
- * This type represents a function that takes a boolean input and produces a byte as result
  * Date: 29/07/17 - 19:57
  */
-public interface FloatToIntFunction extends Function<Float, Integer> {
+public interface FloatToIntFunction extends Function<Integer, Integer> {
 
   int apply(float input);
 
   @Override
-  default Integer apply(Float input){
-    return apply(input.floatValue());
+  default Integer apply(Integer input){
+    return apply(input.intValue());
   }
 }
