+++ src/main/java/ar/com/kfgodel/function/intertype/floats/FloatToObjectFunction.java
@@ -3,15 +3,14 @@ package ar.com.kfgodel.function.intertype.floats;
 import java.util.function.Function;
 
 /**
- * This type represents a function that takes a boolean input and produces a byte as result
  * Date: 29/07/17 - 19:57
  */
-public interface FloatToObjectFunction<T> extends Function<Float, T> {
+public interface FloatToObjectFunction<T> extends Function<Integer, T> {
 
   T apply(float input);
 
   @Override
-  default T apply(Float input){
-    return apply(input.floatValue());
+  default T apply(Integer input){
+    return apply(input.intValue());
   }
 }
